LPS-50967 move interfaces into an SPI package to allow for cleaner Export-Package
