+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/ElasticsearchUpdateDocumentCommandImpl.java
@@ -22,7 +22,8 @@ import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.util.PortalRunMode;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.search.elasticsearch.connection.ElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch.document.ElasticsearchDocumentFactory;
+import com.liferay.portal.search.elasticsearch.spi.ElasticsearchUpdateDocumentCommand;
+import com.liferay.portal.search.elasticsearch.spi.document.ElasticsearchDocumentFactory;
 import com.liferay.portal.search.elasticsearch.util.DocumentTypes;
 import com.liferay.portal.search.elasticsearch.util.LogUtil;
