+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/connection/BaseElasticsearchConnection.java
@@ -16,7 +16,8 @@ package com.liferay.portal.search.elasticsearch.connection;
 
 import com.liferay.portal.kernel.util.PortalRunMode;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.search.elasticsearch.index.IndexFactory;
+import com.liferay.portal.search.elasticsearch.spi.connection.ElasticsearchConnection;
+import com.liferay.portal.search.elasticsearch.spi.index.IndexFactory;
 
 import java.util.concurrent.Future;
