+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/document/DefaultElasticsearchDocumentFactory.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.search.elasticsearch.spi.document.ElasticsearchDocumentFactory;
 
 import java.io.IOException;
