+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/facet/ModifiedFacetProcessor.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.search.elasticsearch.facet;
 
+import com.liferay.portal.search.elasticsearch.spi.facet.FacetProcessor;
+
 import org.osgi.service.component.annotations.Component;
 
 /**
