+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/query/TermRangeQueryTranslatorImpl.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.search.elasticsearch.query;
 
 import com.liferay.portal.kernel.search.TermRangeQuery;
+import com.liferay.portal.search.elasticsearch.spi.query.TermRangeQueryTranslator;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
